Split 'Basin' and 'hopping'
Include explicitly algorithm into the main scope
