+++ votepy/algorithms/__init__.py
@@ -2,3 +2,4 @@ from votepy.algorithms.base_algorithm import BaseAlgorithm
 from votepy.algorithms.brute_force import BruteForce
 from votepy.algorithms.greedy import Greedy
 from votepy.algorithms.p_algorithm import PAlgorithm
+from votepy.algorithms.basin_hopping import BasinHopping
